requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
langchain==0.1.1
langchain-community==0.0.1
langchain-openai==0.0.2
openai==1.3.5
pydantic==2.4.2
requests==2.31.0
validators==0.22.0
